Create testing fixture to test using Elasticsearch spun up by Gradle instead of embedded.
Really sloppy first pass of getting a test fixture that runs ES loaded. This is heavily based on the existing test cluster standup in Elasticsearch, but with most of the functionality torn out of it
Setup and tear down of ES node works.
Upgrading Gradle to 3.3
Add wait task to ensure cluster is started before sending requests.
Pull all gradle distro items instead of only the bin
Gut the EsEmbeddedServer until I can hook up the ES integration tests
Test runs for MR package.
Reorganize the build file
Throw out the commented code, and some old logic from embedded server and local es.
Extract the cluster create into it's own buildscript.
EmbeddedServer to read a ports file from the system properties and add it as the address for the node.
LocalES has been slimmed down to either log that there is no fixture node or to pull and configure the ports file for the fixture node through the embedded server class.
Repos for snapshots in Lucene
LocalEs now creates a template that sets the number of shards created to be 1
Local ES should delete all the indices on the node before executing tests
Should use index_template in LocalEs
Don't stand up a cluster for integration tests if the
integrationTest task is from the root project.
Add ability to create a multi-type index for testing.
If the index does not exist, don't try to delete it.

+++ spark/sql-13/src/main/scala/org/elasticsearch/spark/sql/DefaultSource.scala
@@ -508,7 +508,9 @@ private[sql] case class ElasticsearchRelation(parameters: Map[String, String], @
       cfgCopy.setProperty(ConfigurationOptions.ES_BATCH_SIZE_ENTRIES, "1000")
       cfgCopy.setProperty(ConfigurationOptions.ES_BATCH_SIZE_BYTES, "1mb")
       val rr = new RestRepository(cfgCopy)
-      rr.delete()
+      if (rr.indexExists(false)) {
+        rr.delete()
+      }
       rr.close()
     }
     EsSparkSQL.saveToEs(data, parameters)
